se agrego la entity, el controller y el service de transacciones

+++ src/main/java/com/prueba/demo/dto/TransaccionDto.java
@@ -14,7 +14,7 @@ import java.util.Date;
 public class TransaccionDto {
     private Long id;
     private String tipo;
-    private BigDecimal monto;
+    private Long monto;
     private Date fecha;
     private Cuenta cuentaOrigen;
     private Cuenta cuentaDestino;
